requirements.txt
IMDbPY==2021.4.18
matplotlib==3.4.2
pandas==1.2.4
pyyaml==5.4.1
seaborn==0.11.1
git+https://github.com/asongtoruin/plot_styles#egg=plot_styles